requirements.txt
gnews
feedparser>=6.0.0
google-cloud-storage